collabit-ai/requirements.txt
flask
huggingface_hub
requests
pymongo
python-dotenv
redis
PyJWT
pymysql
flask-cors